Fixes Sente argument passsing in Clojure
The sente function `make-channel-socket!` takes a single configuration map, not a path and a configuration map. Because of this, passing additional configuration options to the `make-channel-socket!` constructor would be ignored. This mistake is silently ignored by Sente because it the signature of the function is `[ & [{ destructuring... }] ]`.

This fixes the issue by removing the `path` argument.